Moved room implementations to a new folder
