+++ LegendOfZelda/Environment/LockedDoor.cs
@@ -1,5 +1,6 @@
 ﻿using LegendOfZelda.Interface;
 using LegendOfZelda.Rooms;
+using LegendOfZelda.Rooms.RoomImplementation;
 using LegendOfZelda.Utility;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
