+++ LegendOfZelda/LegendOfZelda.csproj
@@ -200,8 +200,8 @@
     <Compile Include="GameState\IGameState.cs" />
     <Compile Include="Rooms\Commands\ChangeRoomUpCommand.cs" />
     <Compile Include="Rooms\IRoom.cs" />
-    <Compile Include="Rooms\Room5.cs" />
-    <Compile Include="Rooms\RoomAquamentus.cs" />
+    <Compile Include="Rooms\RoomImplementation\Room5.cs" />
+    <Compile Include="Rooms\RoomImplementation\RoomAquamentus.cs" />
     <Compile Include="Rooms\RoomFactory.cs" />
     <Compile Include="GameLogic\CollisionHandlerDictionary.cs" />
     <Compile Include="GameLogic\CollisionManager.cs" />
@@ -249,7 +249,7 @@
     <Compile Include="Projectile\Sprite\SwordBeamFlyingSprite.cs" />
     <Compile Include="Projectile\SwordBeamFlyingProjectile.cs" />
     <Compile Include="Rooms\CSVReader.cs" />
-    <Compile Include="Rooms\Room.cs" />
+    <Compile Include="Rooms\RoomImplementation\Room.cs" />
     <Compile Include="Projectile\FireballProjectile.cs" />
     <Compile Include="Item\BombItem.cs" />
     <Compile Include="Item\BoomerangItem.cs" />
